Script changes

+++ bin/templates.sh
@@ -1,34 +1,53 @@
-
 #!/bin/bash
 
 ## Builds all the templates using hero, skipping if unchanged
 
-SOURCE="${BASH_SOURCE[0]}"
-while [ -h "$SOURCE" ] ; do SOURCE="$(readlink "$SOURCE")"; done
-dir="$( cd -P "$( dirname "$SOURCE" )/.." && pwd )"
-cd "$dir"
+set -euo pipefail
+dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+cd $dir/..
+
+FORCE="${1:-}"
 
 function tmpl {
-  fsrc="tmp/$1.hashcode"
-  ftgt="tmp/$1.hashcode.tmp"
-
-  if [ ! -d "gen/$1" ]; then
-    rm -f "$fsrc"
-  fi
-
-  find -s "$2" -type f -exec md5sum {} \; | md5sum > "$ftgt"
-
-  if cmp -s "$fsrc" "$ftgt"; then
-    rm "$ftgt"
-  else
-    echo "updating [$2] templates"
+  echo "updating [$2] templates"
+  if test -f "$ftgt"; then
     mv "$ftgt" "$fsrc"
-    rm -rf gen/$1
-    hero -extensions .html,.sql -source "$2" -pkgname $1 -dest gen/$1
+  fi
+  if [ "$1" = "npntemplate" ]; then
+    cd npntemplate
+    rm -rf gen
+    hero -extensions .html,.sql -source "html" -pkgname $1 -dest "gen/npntemplate"
+    cd ..
+  else
+    rm -rf $3
+    hero -extensions .html,.sql -source "$2" -pkgname $1 -dest $3
   fi
 }
 
-tmpl "query" "query/sql"
-tmpl "transcripttemplates" "web/transcript"
-tmpl "templates" "web/templates"
-tmpl "admintemplates" "web/admin"
+function check {
+  fsrc="tmp/$1.hashcode"
+  ftgt="tmp/$1.hashcode.tmp"
+
+  if [ ! -d "$3" ]; then
+    rm -f "$fsrc"
+  fi
+
+  mkdir -p tmp/
+
+  find "$2" -type f -exec md5sum {} \; | md5sum > "$ftgt"
+
+  if cmp -s "$fsrc" "$ftgt"; then
+    if [ "$FORCE" = "force" ]; then
+      tmpl $1 $2 $3
+    else
+      rm "$ftgt"
+    fi
+  else
+    tmpl $1 $2 $3
+  fi
+}
+
+check "query" "query/sql" "gen/query"
+check "transcripttemplates" "web/transcript" "gen/transcripttemplates"
+check "templates" "web/templates" "gen/templates"
+check "admintemplates" "web/admin" "gen/admintemplates"
